PDF Extractor/requirements.txt
numpy==1.26.2
opencv-python==4.8.1.78
pdf2image==1.16.3
Pillow==10.1.0
PyMuPDF==1.23.7
PyMuPDFb==1.23.7
PyPDF2==2.12.1
